feat: 在 IRM 中增加遮挡掩码注意力机制
- MTMBranch 与 CLIMB 图像分支均新增 occlusion_attention MLP
- 拼接全局 cls 与局部 patch 特征，预测遮挡权重 w_t
- 重新校准排序相似度: alpha_new = sim_raw * w_t
- 第一层 Kaiming Normal 初始化，第二层 Xavier Normal + bias=+2.0
- 新增 USE_OCCLUSION_MASK 配置开关，向后兼容
- 同步更新 iLIDS-VID 实验配置文件与数据集加载支持

+++ climb/loss.py
@@ -26,7 +26,7 @@ class CM(autograd.Function):
         inputs, targets = ctx.saved_tensors
         grad_inputs = None
         if ctx.needs_input_grad[0]:
-            grad_inputs = grad_outputs.mm(ctx.features)
+            grad_inputs = grad_outputs.float().mm(ctx.features)
 
         # momentum update
         for x, y in zip(inputs, targets):
@@ -56,7 +56,7 @@ class CM_Hard(autograd.Function):
         inputs, targets = ctx.saved_tensors
         grad_inputs = None
         if ctx.needs_input_grad[0]:
-            grad_inputs = grad_outputs.mm(ctx.features)
+            grad_inputs = grad_outputs.float().mm(ctx.features)
 
         batch_centers = collections.defaultdict(list)
         for instance_feature, index in zip(inputs, targets.tolist()):
@@ -92,7 +92,7 @@ class CM_Mix_mean_hard(autograd.Function):
         nums = len(ctx.features)//2
         grad_inputs = None
         if ctx.needs_input_grad[0]:
-            grad_inputs = grad_outputs.mm(ctx.features)
+            grad_inputs = grad_outputs.float().mm(ctx.features)
         for x, y in zip(inputs, indexes):
             ctx.features[y] = ctx.momentum * ctx.features[y] + (1. - ctx.momentum) * x
             ctx.features[y] /= ctx.features[y].norm()
